上传

+++ file-server/src/main/java/com/jwell/file/server/entity/AttachmentUploadingRecord.java
@@ -41,7 +41,7 @@ public class AttachmentUploadingRecord {
     private Byte attachmentType;
 
     /** 附件大小(kb) */
-    private Long attachmentSize;
+    private Double attachmentSize;
 
     /** 系统编码 例如：1001：设备系统 */
     private String systemCode;
@@ -79,7 +79,7 @@ public class AttachmentUploadingRecord {
 
     }
 
-    public AttachmentUploadingRecord(String attachmentOriginalName, String attachmentName, String fileDirectory, String absoluteAddress, String callOnAddress, String attachmentPostfix, Byte attachmentType, Long attachmentSize, String systemCode, String businessCode, String businessType, String description, String uploaderName, String uploaderId) {
+    public AttachmentUploadingRecord(String attachmentOriginalName, String attachmentName, String fileDirectory, String absoluteAddress, String callOnAddress, String attachmentPostfix, Byte attachmentType, Double attachmentSize, String systemCode, String businessCode, String businessType, String description, String uploaderName, String uploaderId) {
         this.attachmentOriginalName = attachmentOriginalName;
         this.attachmentName = attachmentName;
         this.fileDirectory = fileDirectory;
@@ -162,11 +162,11 @@ public class AttachmentUploadingRecord {
         this.attachmentType = attachmentType;
     }
 
-    public Long getAttachmentSize() {
+    public Double getAttachmentSize() {
         return attachmentSize;
     }
 
-    public void setAttachmentSize(Long attachmentSize) {
+    public void setAttachmentSize(Double attachmentSize) {
         this.attachmentSize = attachmentSize;
     }
